Remove unused base controllers.
With the exception of the import_csv routes, these controllers were
completely unused. This moves the import_csv code into their respective
admin controllers (and correctly auth walled to admins now!) and removes
all of the excess controllers, controller specs, request specs, views,
and routes.
