+++ rails/app/controllers/admin/places_controller.rb
@@ -17,5 +17,17 @@ module Admin
 
     # See https://administrate-prototype.herokuapp.com/customizing_controller_actions
     # for more information
+
+    def import_csv
+      if params[:file].nil?
+        redirect_back(fallback_location: root_path)
+        flash[:error] = "No file was attached!"
+      else
+        filepath = params[:file].read
+        Place.import_csv(filepath)
+        flash[:notice] = "Points were imported successfully!"
+        redirect_back(fallback_location: root_path)
+      end
+    end
   end
 end
