+++ rails/app/controllers/admin/stories_controller.rb
@@ -17,5 +17,16 @@ module Admin
 
     # See https://administrate-prototype.herokuapp.com/customizing_controller_actions
     # for more information
+    def import_csv
+      if params[:file].nil?
+        redirect_back(fallback_location: root_path)
+        flash[:error] = "No file was attached!"
+      else
+        filepath = params[:file].read
+        Story.import_csv(filepath)
+        flash[:notice] = "Stories were imported successfully!"
+        redirect_back(fallback_location: root_path)
+      end
+    end
   end
 end
