--- rails/app/controllers/places_controller.rb
@@ -1,98 +1,0 @@
-class PlacesController < ApplicationController
-  before_action :set_place, only: [:show, :edit, :update, :destroy]
-
-  # GET /places
-  # GET /places.json
-  def index
-    @places = Place.all
-  end
-
-  # GET /places/1
-  # GET /places/1.json
-  def show
-  end
-
-  # GET /places/new
-  def new
-    @place = Place.new
-  end
-
-  # GET /places/1/edit
-  def edit
-  end
-
-  # POST /places
-  # POST /places.json
-  def create
-    @place = Place.new(place_params)
-
-    respond_to do |format|
-      if @place.save
-        format.html { redirect_to @place, notice: 'Place was successfully created.' }
-        format.json { render :show, status: :created, location: @place }
-      else
-        format.html { render :new }
-        format.json { render json: @place.errors, status: :unprocessable_entity }
-      end
-    end
-  end
-
-  # PATCH/PUT /places/1
-  # PATCH/PUT /places/1.json
-  def update
-    respond_to do |format|
-      if @place.update(place_params)
-        format.html { redirect_to @place, notice: 'Place was successfully updated.' }
-        format.json { render :show, status: :ok, location: @place }
-      else
-        format.html { render :edit }
-        format.json { render json: @place.errors, status: :unprocessable_entity }
-      end
-    end
-  end
-
-  # DELETE /places/1
-  # DELETE /places/1.json
-  def destroy
-    @place.destroy
-    respond_to do |format|
-      format.html { redirect_to places_url, notice: 'Place was successfully destroyed.' }
-      format.json { head :no_content }
-    end
-  end
-
-
-  def import_csv
-    if params[:file].nil?
-      redirect_back(fallback_location: root_path)
-      flash[:error] = "No file was attached!"
-    else
-      filepath = params[:file].read
-      Place.import_csv(filepath)
-      flash[:notice] = "Points were imported successfully!"
-      redirect_back(fallback_location: root_path)
-    end
-  end
-
-# delete photo attachment 
-  def delete
-    remove_attachment
-  end
-
-  private
-    # Use callbacks to share common setup or constraints between actions.
-    def set_place
-      @place = Place.find(params[:id])
-    end
-
-    # Never trust parameters from the scary internet, only allow the white list through.
-    def place_params
-      params.require(:place).permit(:name, :type_of_place, :long, :lat, :region)
-    end
-
-    def remove_attachment
-      photo = ActiveStorage::Attachment.find(params[:attachment_id])
-      photo.purge
-      redirect_back(fallback_location: "/")
-    end
-end
