+++ rails/app/views/admin/application/index.html.erb
@@ -67,7 +67,7 @@ It renders the `_table` partial to display details about the resources.
 </section>
 
 <% if ["place", "story", "speaker"].include?(page.resource_name) %>
-  <%= form_tag send("import_csv_#{page.resource_name.pluralize}_path"), multipart: true do %>
+  <%= form_tag send("import_csv_admin_#{page.resource_name.pluralize}_path"), multipart: true do %>
     <div class=form>
       <div class="form-group color-fix">
         <label class="form-title" for="file"><%= t("administrate.file_to_update") %></label>
