+++ rails/config/routes.rb
@@ -2,9 +2,21 @@ Rails.application.routes.draw do
   namespace :admin do
     resources :users
     resources :curriculums
-    resources :speakers
-    resources :stories
-    resources :places
+    resources :speakers do
+      collection do
+        post :import_csv
+      end
+    end
+    resources :stories do
+      collection do
+        post :import_csv
+      end
+    end
+    resources :places do
+      collection do
+        post :import_csv
+      end
+    end
     resources :curriculum_stories
     resources :themes
     resources :media_links
@@ -17,25 +29,9 @@ Rails.application.routes.draw do
     delete '/admin/themes' => 'admin/themes#delete'
 
   scope "(:locale)", locale: Regexp.union(I18n.available_locales.map(&:to_s)) do
-    resources :places do
-      collection do
-        post :import_csv
-      end
-    end
-    resources :stories do
-      collection do
-        post :import_csv
-      end
-    end
-    
     devise_for :users, :controllers => { registrations: 'registrations' }
     # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
     root to: 'welcome#index'
     get 'home', to: 'home#index', as: "home_map"
-    resources :speakers do
-      collection do
-        post :import_csv
-      end
-    end
   end
 end
